Initial checkpoint for py3k support (3.3+ only).
This phase is attempting to be as cosmetic as possible.
There are likely some things that can be done to ensure
type safety between the redis connection which always deals
in bytes types in py3k and our column code which deals in
any python object type.

Added a couple unit tests that seemed relevant during the
process.  However the large number test does not pass with
(cough) large numbers;  It make take some more serious
refactoring to support that or possibly some workarounds in
the base Column code where some type erosion takes place.

The unit tests currently pass for me on Linux CPython 2.7.5
& 3.3.2.