Fix for db cleanup of deleted models in lua mode.
The lua writer would not explicitly know if it needed to remove
the redis backing for a deleted model.  I pass this information
to the lua script so it may perform the same cleanup operation
used by the standard apply_changes() mode.

Corrected session handling in Model.delete().  It was
mistakenly adding itself back to the session cache following a
delete which has the effect of resurrecting it to unsuspecting
users.

Added tests for both fixes identified above.

Fixes #24